+++ docs/changelog.md
@@ -1,7 +1,10 @@
 # Changelog
 
+## Version 0.8 [2024-07-22]
+- [#167](https://github.com/guibranco/pancake/issues/167) - Add ShieldsIo class by [@guibranco](https://github.com/guibranco).
+
 ## Version 0.7 [2024-07-04]
-- [#145](https://github.com/guibranco/pancake/issues/145) - Add Color class by [@gvieiragoulart](https://github.com/gvieiragoulart)
+- [#145](https://github.com/guibranco/pancake/issues/145) - Add Color class by [@gvieiragoulart](https://github.com/gvieiragoulart).
   
 ## Version 0.6 [2024-04-21]